Switched correlation id to smaller sized number for better Lua compatiblity.

+++ common/src/schemas/event.rs
@@ -10,7 +10,7 @@ use schemas::consistency::{
 #[derive(Serialize, Deserialize, Clone)]
 pub struct Event {
     pub consistency: Consistency,
-    pub correlation_id: usize,
+    pub correlation_id: u32,
     pub data: Value,
     pub event_type: String,
     // We don't want this field going to Kafka.
